+++ packages/explorer-ui/components/misc/ChainInfo.tsx
@@ -1,8 +1,8 @@
-import { QuestionMarkCircleIcon } from '@heroicons/react/outline'
-import { getChainUrl, getExplorerTxUrl } from '@urls'
+import Link from 'next/link'
 import Image from 'next/image'
 import { CHAINS } from 'synapse-constants'
-import Link from 'next/link'
+import { QuestionMarkCircleIcon } from '@heroicons/react/outline'
+import { getChainUrl, getExplorerTxUrl } from '@urls'
 
 const CHAINS_BY_ID = CHAINS.CHAINS_BY_ID
 
@@ -17,16 +17,14 @@ interface ChainInfoProps {
   className?: string
 }
 
-export function ChainInfo({
+export const ChainInfo = ({
   chainId,
   imgClassName = 'w-4 h-4 rounded-full',
-  linkClassName = 'float-right text-white transition ease-out hover:text-[#8FEBFF] px-1.5  rounded-md ease-in-out bg-[#191919]',
   textClassName = 'pl-1 whitespace-nowrap text-sm text-white',
   txHash,
   useExplorerLink = false,
-  noLink = false,
   className = '',
-}: ChainInfoProps) {
+}: ChainInfoProps) => {
   const chain = CHAINS_BY_ID[chainId]
   let link = ''
 
@@ -39,16 +37,6 @@ export function ChainInfo({
   }
 
   if (chain) {
-    const content = (
-      <>
-        <Image
-          className={`inline rounded-full ${imgClassName}`}
-          src={chain?.chainImg}
-          alt={chain?.name}
-        />
-        <p className={textClassName}>{chain.name}</p>
-      </>
-    )
     return (
       <div className="relative w-full">
         <Link href={link} passHref legacyBehavior>
@@ -58,7 +46,11 @@ export function ChainInfo({
               src={chain?.chainImg}
               alt={chain?.name}
             />
-            <p className={`${textClassName} group-hover:text-[#8FEBFF] transition-colors duration-200`}>{chain.name}</p>
+            <p
+              className={`${textClassName} group-hover:text-[#8FEBFF] transition-colors duration-200`}
+            >
+              {chain.name}
+            </p>
           </div>
         </Link>
       </div>