+++ packages/explorer-ui/utils/checksum.ts
@@ -1,6 +1,6 @@
 const keccak = require('keccak')
 
-export function checksumAddress(address, chainId = null) {
+export const checksumAddress = (address, chainId = null) => {
   if (typeof address !== 'string') {
     return ''
   }
@@ -25,7 +25,7 @@ export function checksumAddress(address, chainId = null) {
   return checksumAddress
 }
 
-export function checkAddressChecksum(address) {
+export const checkAddressChecksum = (address) => {
   const stripAddress = stripHexPrefix(address).toLowerCase()
   const prefix = '0x'
   const keccakHash = keccak('keccak256')
@@ -44,6 +44,6 @@ export function checkAddressChecksum(address) {
   return true
 }
 
-function stripHexPrefix(value) {
+const stripHexPrefix = (value) => {
   return value.slice(0, 2) === '0x' ? value.slice(2) : value
 }