+++ services/rfq/relayer/relapi/server.go
@@ -59,7 +59,7 @@ func NewRelayerAPI(
 	}
 
 	chains := make(map[uint32]*chain.Chain)
-	for chainID, chainCfg := range cfg.Chains {
+	for chainID := range cfg.Chains {
 		chainClient, err := omniRPCClient.GetChainClient(ctx, chainID)
 		if err != nil {
 			return nil, fmt.Errorf("could not create omnirpc client: %w", err)
@@ -80,7 +80,7 @@ func NewRelayerAPI(
 		if err != nil {
 			return nil, fmt.Errorf("could not get chain listener: %w", err)
 		}
-		chains[uint32(chainID)], err = chain.NewChain(ctx, chainClient, common.HexToAddress(chainCfg.RFQAddress), chainListener, submitter)
+		chains[uint32(chainID)], err = chain.NewChain(ctx, cfg, chainClient, chainListener, submitter)
 		if err != nil {
 			return nil, fmt.Errorf("could not create chain: %w", err)
 		}
@@ -112,7 +112,7 @@ func (r *RelayerAPIServer) Run(ctx context.Context) error {
 	engine := ginhelper.New(logger)
 	// default tracing middleware
 	engine.Use(r.handler.Gin()...)
-	h := NewHandler(r.db, r.chains, r.cfg, r.submitter)
+	h := NewHandler(r.handler, r.db, r.chains, r.cfg, r.submitter)
 
 	// Assign GET routes
 	engine.GET(getHealthRoute, h.GetHealth)